backend/requirements.txt
alembic==1.17.2
bcrypt==5.0.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.11
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
more-itertools==10.8.0
msgspec==0.19.0
numpy==2.3.5
pandas==2.3.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlacodegen==3.1.1
SQLAlchemy==2.0.41
typeguard==4.4.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
flask-sqlalchemy
